incomr check

+++ app/Modules/AccountManagement/AccountCategory/Actions/Delete.php
@@ -13,6 +13,11 @@ class Delete
             if (!$data = self::$model::find($id)) {
                 return messageResponse('Data not found...', 404, 'error');
             }
+
+            if ($data->is_default == 1) {
+                return messageResponse('Sorry you can not deleted this item', 404, 'error');
+            }
+
             $accountCategory = self::$accountLogModel::find($id);
             if (!$accountCategory) {
                 $data->delete();
